[ao] Adds framework for InputWeightEqualization Detector
Summary: This adds the framework (method signatures and descriptors) for
the InputWeightEqualization Detector. There is no code implemenation yet
so the test suite for this is a simple pass. This Detector will be used
to determine whether input weight equalization should be recommended.

Test Plan: python test/test_quantization.py TestFxDetectInputWeightEqualization

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 3e23ca9b5d64d9c82be001b2610d04dbc28503e0
Pull Request resolved: https://github.com/pytorch/pytorch/pull/79916